requirements.txt
faker>=24.0.0
lancedb>=0.10.0
sentence-transformers>=3.0.0
numpy>=1.26.0
pandas>=2.2.0
mcp>=1.0.0
pyjwt>=2.8.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
